Move App.tsx folder into "app" folder for global configurations
